Session based auth

+++ api/server/posts.go
@@ -3,9 +3,22 @@ package server
 import (
 	"net/http"
 
-	"../models"
 	"github.com/gin-gonic/gin"
+	"github.com/mcavoyk/quirk/models"
 )
+
+type Post struct {
+	ParentID   string
+	Title      string
+	Content    interface{}
+	AccessType string
+	Latitude   float64
+	Longitude  float64
+}
+
+func convertPost(src *Post) *models.Post {
+	return &models.Post{}
+}
 
 func (env *Env) PostGet(c *gin.Context) {
 	id := c.Param("id")
@@ -29,6 +42,13 @@ func (env *Env) PostDelete(c *gin.Context) {
 		return
 	}
 
+	currentPost := env.DB.GetPost(id)
+	if currentPost.User != c.GetString(UserContext) {
+		c.JSON(http.StatusForbidden, gin.H{
+			"Invalid": "Post ID can not be empty",
+		})
+		return
+	}
 	env.DB.DeletePost(id)
 	c.JSON(http.StatusOK, http.StatusText(http.StatusOK))
 }
@@ -39,16 +59,19 @@ func (env *Env) PostPatch(c *gin.Context) {
 		return
 	}
 
-	env.DB.InsertPost(post)
+	env.DB.UpdatePost(post)
 	c.JSON(http.StatusOK, http.StatusText(http.StatusOK))
 }
 
 func (env *Env) PostPost(c *gin.Context) {
-	post := &models.Post{}
+	post := &Post{}
 	if err := c.Bind(post); err != nil {
 		return
 	}
 
-	env.DB.UpdatePost(post)
+	newPost := convertPost(post)
+	newPost.User = c.GetString(UserContext)
+
+	env.DB.InsertPost(newPost)
 	c.JSON(http.StatusOK, http.StatusText(http.StatusOK))
 }
